Switch mine to claim v2
