+++ programs/quarry-mine/src/instructions/mod.rs
@@ -1,5 +1,7 @@
+pub mod claim_rewards;
 pub mod claim_rewards_v2;
 pub mod rescue_tokens;
 
+pub use claim_rewards::*;
 pub use claim_rewards_v2::*;
 pub use rescue_tokens::*;
